src/cloud-function/requirements.txt
Flask
google-cloud-pubsub
functions-framework==3.*